requirements.txt
requests>=2.31.0
watchdog>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
pyyaml>=6.0
python-dotenv>=1.0.0
GitPython>=3.1.0
jinja2>=3.1.0
plotly>=5.15.0